funcionalidad de anadir usuarios al team remover y listar los miembros del proyecto

+++ src/controllers/ProjectController.ts
@@ -6,7 +6,9 @@ export class ProjectController {
 
   static createdProjects = async (req: Request, res: Response) => {
     const project = new Project(req.body);
-
+    
+    //Asignar un manager 
+    project.manager = req.user.id // este req.user.id se lo estoy pasando desde el middleware de autenticate
     try {
       await project.save(); // lo puedo hacer de otra manera como await Project.create(req.body)
       res.send("Proyecto creado correctamente");
@@ -17,7 +19,11 @@ export class ProjectController {
 
   static getAllProjects = async (req: Request, res: Response) => {
     try {
-      const projects = await Project.find();
+      const projects = await Project.find({
+        $or:[
+          {manager: {$in: req.user.id }}
+        ]
+      });
       res.json(projects);
     } catch (error) {
       console.log(error);
@@ -32,6 +38,12 @@ export class ProjectController {
         const error = new Error("Proyecto no encontrado");
         return res.status(404).json({ error: error.message });
       }
+
+      if(project.manager.toString() !== req.user.id){
+        const error = new Error("Accion no valida,no tienes el permiso necesario.");
+        return res.status(401).json({ error: error.message });
+      }
+
       res.json(project);
     } catch (error) {
       console.log(error);
@@ -46,6 +58,11 @@ export class ProjectController {
       if (!project) {
         const error = new Error("Proyecto no encontrado");
         return res.status(404).json({ error: error.message });
+      }
+
+      if(project.manager.toString() !== req.user.id){
+        const error = new Error("Solo el administrador puede actualizar el proyecto");
+        return res.status(401).json({ error: error.message });
       }
       project.clientName = req.body.clientName
       project.projectName = req.body.projectName
@@ -65,6 +82,11 @@ export class ProjectController {
       if (!project) {
         const error = new Error("Proyecto no encontrado");
         return res.status(404).json({ error: error.message });
+      }
+
+      if(project.manager.toString() !== req.user.id){
+        const error = new Error("Solo el administrador puede eliminar el proyecto");
+        return res.status(401).json({ error: error.message });
       }
 
       await project.deleteOne(); // este metodo elimina una instacia
